% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_hex.R
\name{cast_hex_grid}
\alias{cast_hex_grid}
\title{Create a hexagon grid for an input set of plots}
\usage{
cast_hex_grid(plots, side_to_side = 5187, attributes = c("z_1"))
}
\value{
A SpatialPolygonsDataFrame of hexagons with attributes appended
}
\description{
Create a hexagon grid for an input set of plots
}
