% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{compare_estimators}
\alias{compare_estimators}
\title{Using subsampling, assess a series of variance estimators.}
\usage{
compare_estimators(sys_frame, a_vec, estimators)
}
\description{
Using subsampling, assess a series of variance estimators.
}
