% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhoods.R
\name{get_mat_hex_neighborhoods}
\alias{get_mat_hex_neighborhoods}
\title{The Matern estimator does not use 'centers' per se
but rather 'anchors'. Also, a set of contrasts is needed
for each neighborhood.}
\usage{
get_mat_hex_neighborhoods(anchors, contrasts = c(1, 1, -1, -1))
}
\description{
The Matern estimator does not use 'centers' per se
but rather 'anchors'. Also, a set of contrasts is needed
for each neighborhood.
}
