% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HexFrame.R
\name{index_hex_polys}
\alias{index_hex_polys}
\title{Creates a dataframe of hexagon indices for use in \code{subsample_hex}}
\usage{
index_hex_polys(hex_polys)
}
\arguments{
\item{hex_polys}{A SpatialPolygonsDataFrame of a hexagon grid.}
}
\value{
A dataframe with two columns, r and c, that correspond to row and column indices
of the hexagon grid.
}
\description{
Creates a dataframe of hexagon indices for use in \code{subsample_hex}
}
