% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{index_sq}
\alias{index_sq}
\title{Creates a dataframe of hexagon indices for use in \code{subsample_sq}.}
\usage{
index_sq(square_polys)
}
\arguments{
\item{square_polys}{A SpatialPolygonsDataFrame of a square grid.}
}
\value{
A dataframe with two columns, r and c, that correspond to row and column indices
of the hexagon grid.
}
\description{
Creates a dataframe of hexagon indices for use in \code{subsample_sq}.
}
