% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_methods.R
\name{pop_var}
\alias{pop_var}
\title{Calculate the variance using a denominator of n
instead of n-1}
\usage{
pop_var(z)
}
\description{
Calculate the variance using a denominator of n
instead of n-1
}
\keyword{internal}
