% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_methods.R
\name{so_att}
\alias{so_att}
\title{Computes the Stevens and Olsen estimator for a specific attribute.
Used internally with var_so function.}
\usage{
so_att(att, att_df, pi_i, wt)
}
\description{
Computes the Stevens and Olsen estimator for a specific attribute.
Used internally with var_so function.
}
