% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HexFrame.R
\name{subsample}
\alias{subsample}
\title{Create a systematic sample from an existing \code{SysFrame}}
\usage{
subsample(object, start_pos, a)
}
\arguments{
\item{object}{A \code{SysFrame} from which to sample}

\item{start_pos}{A numeric vector of two elements indicating the starting position}

\item{a}{The sampling interval}
}
\value{
A new \code{SysFrame} object that contains only the sampled elements
}
\description{
Creates a systematic sample from a \code{SysFrame} for a specified starting position
and sampling interval. Typically used to conduct variance assessments, e.g. as a part of
\code{compare_estimators()}.
}
