% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample_hex_ix_compact}
\alias{subsample_hex_ix_compact}
\title{There is a way to subsample a set of hexagonal
indices such that we obtain a compact set of
neighborhoods. Implemented here.}
\usage{
subsample_hex_ix_compact(ix)
}
\description{
There is a way to subsample a set of hexagonal
indices such that we obtain a compact set of
neighborhoods. Implemented here.
}
