% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample_hex_ix_ov}
\alias{subsample_hex_ix_ov}
\title{Creates a set of indices such that their nearest
neighbors create the overlapping structure}
\usage{
subsample_hex_ix_ov(ix)
}
\description{
Creates a set of indices such that their nearest
neighbors create the overlapping structure
}
