% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample_starts}
\alias{subsample_starts}
\title{For a given sampling interval, return the set of all possible starting positions}
\usage{
subsample_starts(sys_frame, a)
}
\arguments{
\item{a}{The sampling interval}
}
\value{
A dataframe of starting positions for all possible systematic samples
for the interval a.
}
\description{
For a given sampling interval, return the set of all possible starting positions
}
