% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{translate_hex_ix}
\alias{translate_hex_ix}
\title{For a given set of (subsampled) hexagonal indices, translate
to the base coordinate system.}
\usage{
translate_hex_ix(hex_ix, a)
}
\description{
For a given set of (subsampled) hexagonal indices, translate
to the base coordinate system.
}
