% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_methods.R
\name{var_so}
\alias{var_so}
\title{Computes the variance estimate of a systematic
sample using the Stevens and Olsen (2003) local mean estimator.}
\usage{
var_so(sys_frame, ...)
}
\description{
Computes the variance estimate of a systematic
sample using the Stevens and Olsen (2003) local mean estimator.
}
