% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarEst.R
\name{VarMAT}
\alias{VarMAT}
\title{Matern (1984) local variance estimator}
\usage{
VarMAT(fpc = FALSE, diagnostic = TRUE, nbh = "mat")
}
\arguments{
\item{fpc}{If TRUE, finite population correction factor will be applied.}

\item{diagnostic}{If TRUE, diagnostic information, i.e. a VarOut, object will be returned.}

\item{nbh}{The neighborhood structure. For \code{HexFrame} one of either \code{mat} or \code{hex}, for \code{RectFrame} only
\code{mat} is supported and reflects the original structure implemented by Matern.}
}
\value{
Returns a function with the specified arguments that can be called on a
SysFrame.
}
\description{
Constructor for a local variance estimator introduced by Matern in [] and [].
}
