% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarEst.R
\name{VarSO}
\alias{VarSO}
\title{Neighborhood-based estimator from \insertCite{stevens_variance_2003;textual}{spsys}}
\usage{
VarSO(fpc = FALSE, diagnostic = FALSE, coord_cols = NA, nbh = 4)
}
\arguments{
\item{fpc}{If TRUE, finite population correction factor will be applied.}

\item{diagnostic}{If TRUE, diagnostic information, i.e. a VarOut, object will be returned.}

\item{coord_cols}{A character vector of the names of the coordinates in the \code{sys_frame} that
contain the x and y coordinates. The x column name should be listed first.}

\item{nbh}{The number of neighbors to use in the construction of local neighborhoods.}
}
\value{
Returns a function with the specified arguments that can be called on a
SysFrame.
}
\description{
Constructor for a local variance estimator introduced in \insertCite{stevens_variance_2003;textual}{spsys}.
This is a wrapper
for an implementation in the \code{spsurvey} package. A column named \code{pi_i} must be present in the dataframe
contained in the \code{@data} slot that specifies the first order inclusion probabilities of each element in the sample.
}
\references{
{
\insertAllCited{}
}
}
