% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarEst.R
\name{VarSRS}
\alias{VarSRS}
\title{Variance estimator assuming simple random sampling}
\usage{
VarSRS(fpc = FALSE, diagnostic = FALSE)
}
\arguments{
\item{fpc}{If TRUE, finite population correction factor will be applied.}

\item{diagnostic}{If TRUE, diagnostic information, i.e. a VarOut, object will be returned.}
}
\value{
Returns a function with the specified arguments that can be called on a
SysFrame.
}
\description{
Constructor for the variance estimator assuming simple random sampling.
}
\examples{
v_srs <- VarSRS(fpc=TRUE)
v_srs(hex_frame)
}
