% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhoods.R
\name{get_hex_neighborhoods}
\alias{get_hex_neighborhoods}
\title{For some input set of neighborhood center points
return all immediate hexagonal neighbors.
TODO could make this Hex/Rect method
TODO the default contrasts is a bit messy...}
\usage{
get_hex_neighborhoods(centers, contrasts = c(1, 1, 1, 0, -1, -1, -1))
}
\description{
For some input set of neighborhood center points
return all immediate hexagonal neighbors.
TODO could make this Hex/Rect method
TODO the default contrasts is a bit messy...
}
