% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HexFrame.R
\name{subsample}
\alias{subsample}
\title{Subsamples a dataframe of hexagonal indices produced by the \code{index_hex} function.
For some specified starting index and sampling interval, subsamples a hexagonal index
set such that the spatial structure is preserved, i.e. subsamples also retain a
hexagonal structure.}
\usage{
subsample(object, start_pos, a, standardize = TRUE)
}
\arguments{
\item{start_pos}{the starting position}

\item{a}{The order of the subset: 2 represents sampling every other}

\item{hex_ix}{A dataframe of hex indices from the \code{index_hex} function.}
}
\value{
A dataframe of row and column indices that have been sampled.
}
\description{
Subsamples a dataframe of hexagonal indices produced by the \code{index_hex} function.
For some specified starting index and sampling interval, subsamples a hexagonal index
set such that the spatial structure is preserved, i.e. subsamples also retain a
hexagonal structure.
}
