% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample_hex_ix_compact}
\alias{subsample_hex_ix_compact}
\title{Retrieves a subset of indices that correspond to the center of a compact set
of hexagonal neighborhoods}
\usage{
subsample_hex_ix_compact(ix)
}
\arguments{
\item{ix}{A dataframe of hexagonal indices}
}
\value{
A dataframe of hexagonal neigborhood centers
}
\description{
Retrieves a subset of indices that correspond to the center of a compact set
of hexagonal neighborhoods
}
\keyword{internal}
