% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_methods.R
\name{var_so}
\alias{var_so}
\title{Internal function used to compute the Stevens and Olsen (2003) variance estimate.}
\usage{
var_so(sys_frame, ...)
}
\arguments{
\item{sys_frame}{A \code{SysFrame} object}

\item{fpc}{If TRUE, use the finite population correction factor}

\item{diagnostic}{If TRUE return diagnostic information, else return only the variance estimate}

\item{coord_cols}{A named vector of coordinate columns}

\item{nbh}{An integer representing the number of neighbors to use in the variance estimate}
}
\description{
Internal function used to compute the Stevens and Olsen (2003) variance estimate.
}
\keyword{internal}
