% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarEst.R
\name{VarMAT}
\alias{VarMAT}
\title{\insertCite{matern_spatial_1986;textual}{spsys} local variance estimator}
\usage{
VarMAT(fpc = FALSE, diagnostic = FALSE, nbh = "par")
}
\arguments{
\item{fpc}{If TRUE, finite population correction factor will be applied.}

\item{diagnostic}{If TRUE, diagnostic information, i.e. a VarOut, object will be returned.}

\item{nbh}{The neighborhood structure. For \code{HexFrame} one of either \code{'par'} or \code{'hex'}, for \code{RectFrame} only
\code{par} is supported and reflects the original structure implemented by Matèrn}
}
\value{
Returns a function with the specified arguments that can be called on a `SysFrame``
}
\description{
Constructor for a local variance estimator introduced in \insertCite{matern_spatial_1986;textual}{spsys}.
}
\references{
{
\insertAllCited{}
}
}
