% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarEst.R
\name{VarNON}
\alias{VarNON}
\title{Non-overlapping stratified random sampling estimator}
\usage{
VarNON(fpc = FALSE, diagnostic = FALSE, nbh = "par")
}
\arguments{
\item{fpc}{If TRUE, finite population correction factor will be applied}

\item{diagnostic}{If TRUE, diagnostic information, i.e. a VarOut, object will be returned}

\item{nbh}{The neighborhood structure. For \code{HexFrame} one of either \code{tri} \code{par} or \code{hex}
For \code{RectFrame} only \code{par} is supported.}
}
\value{
Returns a function with the specified arguments that can be called on a \code{SysFrame}
}
\description{
Constructor for a local variance estimator that aggregates variance estimates
from non-overlapping strata. An example of this estimator is given in \insertCite{aune-lundberg_comparison_2014;textual}{spsys}.
}
\references{
{
\insertAllCited{}
}
}
